Package diffusion2d as rongchzg_diffusion2d
