+++ rongchzg_diffusion2d.egg-info/requires.txt
@@ -1,0 +1,2 @@
+numpy
+matplotlib
