+++ rongchzg_diffusion2d/__init__.py
@@ -1,0 +1,5 @@
+"""
+Package for solving the two-dimensional diffusion equation.
+"""
+
+from .diffusion2d import solve
